moving c8y login to separate package
