+++ go.mod
@@ -8,11 +8,11 @@ require (
 	github.com/howeyc/gopass v0.0.0-20190910152052-7cb4b85ec19c
 	github.com/karrick/tparse/v2 v2.8.2
 	github.com/manifoldco/promptui v0.7.0
-	github.com/mdp/qrterminal/v3 v3.0.0
+	github.com/mdp/qrterminal v1.0.1
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/op/go-logging v0.0.0-20160315200505-970db520ece7
 	github.com/pkg/errors v0.9.1
-	github.com/pquerna/otp v1.2.0 // indirect
+	github.com/pquerna/otp v1.2.0
 	github.com/reubenmiller/go-c8y v0.8.1-0.20201024115906-9e1f3ed4442f
 	github.com/sethvargo/go-password v0.2.0
 	github.com/spf13/cobra v0.0.5
