+++ tools/PSc8y/Public-manual/Invoke-ClientLogin.ps1
@@ -2,7 +2,10 @@ Function Invoke-ClientLogin {
     [cmdletbinding()]
     Param(
         # Two Factor Authentication code
-        [string] $TFACode
+        [string] $TFACode,
+
+        # Clear existing cookies (if present)
+        [switch] $Clear
     )
     Process {
         $c8yBinary = Get-ClientBinary
@@ -10,6 +13,14 @@ Function Invoke-ClientLogin {
         $cliArgs = New-Object System.Collections.ArrayList
 
         $null = $cliArgs.AddRange(@("sessions", "login"))
+
+        if ($TFACode) {
+            $null = $cliArgs.AddRange(@("--tfaCode", $TFACode))
+        }
+
+        if ($Clear) {
+            $null = $cliArgs.AddRange(@("--clear"))
+        }
 
         if ($VerbosePreference) {
             $cliArgs.Add("--verbose")
